+++ modules/core-config/main.tf
@@ -160,6 +160,7 @@ module "fluent_bit" {
   aggregator_forward_port = var.core_services_config.fluent_bit_aggregator.enabled ? module.fluent_bit_aggregator[0].forward_port : module.fluentd[0].forward_port
 
   multiline_parsers = var.experimental.fluent_bit_collector_multiline_parsers
+  parsers           = var.experimental.fluent_bit_collector_parsers
 
   timeouts = var.timeouts