+++ CHANGELOG.md
@@ -33,6 +33,7 @@ All clusters created with a module version older than `v1.0.0-beta.10` need to b
 - Removed -->
 
 ## [v1.18.0] - UNRELEASED
+- Added experimental feature of custom fluent bit single line parser ([#1235](https://github.com/LexisNexis-RBA/rsg-terraform-azurerm-aks/pull/1235) [@james-alford-ln](https://github.com/james-alford-ln)
 
 ### Highlights